third_party/blink/web_tests/external/wpt/html-aam/dir-role.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] el-dir
  assert_equals: <dir data-testname="el-dir" data-expectedrole="list" class="ex"><li>x</li><li>x</li></dir> expected "list" but got "none"
Harness: the test ran to completion.
